pyproject.toml
[project]
name = "erpnext_enhancements"
authors = [
    { name = "Sapphire Fountains", email = "info@sapphirefountains.com"}
]
description = "Customizations and enhancements to ERPNext."
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "google-generativeai",
    "google-analytics-data",
    # "frappe~=16.0.0" # Installed and managed by bench.
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

[tool.bench]
# frappe-dependencies must be a table, not a list.
# Strict versioning is required (Lower and Upper bound).
[tool.bench.frappe-dependencies]
frappe = ">=16.0.0,<17.0.0"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
faker = "~=18.0.0"
responses = "~=0.23.0"

[tool.ruff]
line-length = 110
target-version = "py310"

[tool.ruff.lint]
select = [
    "F",
    "E",
    "W",
    "I",
    "UP",
    "B",
    "RUF",
]
ignore = [
    "B017", # assertRaises(Exception) - should be more specific
    "B018", # useless expression, not assigned to anything
    "B023", # function doesn't bind loop variable - will have last iteration's value
    "B904", # raise inside except without from
    "E101", # indentation contains mixed spaces and tabs
    "E402", # module level import not at top of file
    "E501", # line too long
    "E741", # ambiguous variable name
    "F401", # "unused" imports
    "F403", # can't detect undefined names from * import
    "F405", # can't detect undefined names from * import
    "F722", # syntax error in forward type annotation
    "W191", # indentation contains tabs
]
typing-modules = ["frappe.types.DF"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
docstring-code-format = true
